Codigo

+++ Tradiciones/headers/Window.h
@@ -1,7 +1,7 @@
 /**
 * @file Window.h
 * @class Window
-* 
+*
 * @brief Inicializa el contexto de opengl
 */
 #pragma once
@@ -12,21 +12,19 @@
 class Window
 {
 public:
-
 	Window();
 	Window(GLint windowWidth, GLint windowHeight);
 	int Initialise();
-
-	GLfloat getBufferWidth() { return (float)bufferWidth; }
-	GLfloat getBufferHeight() { return (float)bufferHeight; }
+	GLfloat getBufferWidth() { return bufferWidth; }
+	GLfloat getBufferHeight() { return bufferHeight; }
 	GLfloat getXChange();
 	GLfloat getYChange();
 	bool getShouldClose() {
 		return  glfwWindowShouldClose(mainWindow);}
 	bool* getsKeys() { return keys; }
 	void swapBuffers() { return glfwSwapBuffers(mainWindow); }
+	
 	~Window();
-
 private: 
 	GLFWwindow *mainWindow;
 	GLint width, height;
@@ -40,5 +38,6 @@ private:
 	bool mouseFirstMoved;
 	static void ManejaTeclado(GLFWwindow* window, int key, int code, int action, int mode);
 	static void ManejaMouse(GLFWwindow* window, double xPos, double yPos);
+
 };
